Tab component
I realized I just jumped right in with trunk-based development here, because it works great for a solo developer working alone (which is what's happening here), but also mob programming. But it could look like the result of naïvety, so from here on I'm going to follow Git flow and make pull requests, the way you might for an open source project with many contributors. This provides a great opportunity for code review in the pull requests when you're working in an environment like that, but for a project like this, it's going through the motions primarily to demonstrate that I'm familiar with how it's done.